requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.12.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-groq==0.2.1
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.5
numpy==1.26.4
ollama==0.3.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.35
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
